test_push_notifications.py
#!/usr/bin/env python
"""
Test script for push notification functionality in Smart Tracker
"""
import os
import sys
import django
from django.conf import settings
from django.test.utils import get_runner

# Add the project directory to the path
sys.path.insert(0, os.path.dirname(__file__))

# Configure Django settings
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'expirytracker.settings')

# Setup Django
django.setup()

def test_push_notifications():
    """Test push notification functionality"""
    print("🔔 Testing Push Notification Implementation")
    print("=" * 50)

    # Test 1: Check Django system check
    print("✅ Test 1: Django System Check")
    from django.core.management import execute_from_command_line
    try:
        execute_from_command_line(['manage.py', 'check'])
        print("   ✓ Django system check passed")
    except Exception as e:
        print(f"   ✗ Django system check failed: {e}")
        return False

    # Test 2: Check VAPID settings
    print("\n✅ Test 2: VAPID Configuration")
    try:
        from django.conf import settings
        vapid_private = getattr(settings, 'VAPID_PRIVATE_KEY', None)
        vapid_public = getattr(settings, 'VAPID_PUBLIC_KEY', None)
        vapid_claims = getattr(settings, 'VAPID_CLAIMS', None)

        if vapid_private and vapid_public and vapid_claims:
            print("   ✓ VAPID keys configured")
        else:
            print("   ✗ VAPID keys not properly configured")
            return False
    except Exception as e:
        print(f"   ✗ VAPID configuration error: {e}")
        return False

    # Test 3: Check models
    print("\n✅ Test 3: Database Models")
    try:
        from tracker.models import PushSubscription, UserProfile
        from django.contrib.auth.models import User

        # Check if PushSubscription model exists
        push_sub = PushSubscription.objects.all()
        print("   ✓ PushSubscription model accessible")

        # Check if UserProfile has push_reminders_enabled field
        user_profile_fields = [field.name for field in UserProfile._meta.fields]
        if 'push_reminders_enabled' in user_profile_fields:
            print("   ✓ UserProfile.push_reminders_enabled field exists")
        else:
            print("   ✗ UserProfile.push_reminders_enabled field missing")
            return False

    except Exception as e:
        print(f"   ✗ Model test failed: {e}")
        return False

    # Test 4: Check API views
    print("\n✅ Test 4: API Views")
    try:
        from tracker.api_views import vapid_public_key, subscribe_push, unsubscribe_push, send_push_notification
        print("   ✓ Push notification API views imported successfully")
    except ImportError as e:
        print(f"   ✗ API views import failed: {e}")
        return False

    # Test 5: Check Celery task
    print("\n✅ Test 5: Celery Task")
    try:
        from tracker.tasks import send_expiry_reminders
        print("   ✓ Celery task imported successfully")
    except ImportError as e:
        print(f"   ✗ Celery task import failed: {e}")
        return False

    # Test 6: Check dependencies
    print("\n✅ Test 6: Dependencies")
    try:
        import pywebpush
        print("   ✓ pywebpush installed")
    except ImportError:
        print("   ✗ pywebpush not installed")
        return False


    # Test 7: Check URLs
    print("\n✅ Test 7: URL Configuration")
    try:
        from django.urls import reverse
        # Test if notification URLs are configured
        from tracker.urls import api_urlpatterns
        notification_urls = [pattern.pattern._route for pattern in api_urlpatterns if 'notifications' in str(pattern.pattern)]
        if notification_urls:
            print(f"   ✓ Notification URLs configured: {len(notification_urls)} endpoints")
        else:
            print("   ✗ No notification URLs found")
            return False
    except Exception as e:
        print(f"   ✗ URL configuration test failed: {e}")
        return False

    print("\n🎉 All tests passed! Push notification implementation is ready.")
    print("\n📋 Summary:")
    print("   • Database models: PushSubscription + UserProfile enhancement")
    print("   • API endpoints: 3 notification endpoints")
    print("   • Celery task: Enhanced with push notifications")
    print("   • Dependencies: pywebpush installed")
    print("   • Configuration: VAPID keys set up")
    print("\n🚀 Ready for production deployment!")

    return True

if __name__ == '__main__':
    success = test_push_notifications()
    sys.exit(0 if success else 1)
